+++ src/components/Appbar.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import SigninButton from './SigninButton';
 
+// CSS Formatting the Signin Button
+
 const Appbar = () => {
   return (
     <header className="flex gap-4 p-4 bg-gradient-to-b from-white to-gray-200 shadow">
