+++ src/components/Providers.tsx
@@ -2,6 +2,7 @@
 import { SessionProvider } from 'next-auth/react';
 import React, { ReactNode } from 'react';
 
+// Returning the next auth provider
 interface Props {
   children: ReactNode;
 }
